Move package_query libs to depscan.lib, update pyproject.toml.
