+++ test/test_pkg_query.py
@@ -5,12 +5,12 @@ import pytest
 
 from depscan.lib import config as config
 from depscan.lib.bom import get_pkg_list
-from depscan.package_query.pkg_query import (
+from depscan.lib.package_query.pkg_query import (
     calculate_risk_score,
     get_category_score, )
-from depscan.package_query.metadata import npm_metadata, pypi_metadata
-from depscan.package_query.pypi_pkg import pypi_pkg_risk
-from depscan.package_query.npm_pkg import npm_pkg_risk
+from depscan.lib.package_query.metadata import npm_metadata, pypi_metadata
+from depscan.lib.package_query.pypi_pkg import pypi_pkg_risk
+from depscan.lib.package_query.npm_pkg import npm_pkg_risk
 
 
 def test_risk_scores():
